Refactored version control
