+++ src/main/java/seedu/academydirectory/logic/commands/Command.java
@@ -1,7 +1,7 @@
 package seedu.academydirectory.logic.commands;
 
 import seedu.academydirectory.logic.commands.exceptions.CommandException;
-import seedu.academydirectory.model.Model;
+import seedu.academydirectory.model.VersionedModel;
 
 /**
  * Represents a command with hidden internal logic and the ability to be executed.
@@ -15,6 +15,6 @@ public abstract class Command {
      * @return feedback message of the operation result for display
      * @throws CommandException If an error occurs during command execution.
      */
-    public abstract CommandResult execute(Model model) throws CommandException;
+    public abstract CommandResult execute(VersionedModel model) throws CommandException;
 
 }
