+++ src/main/java/seedu/academydirectory/logic/commands/ExitCommand.java
@@ -1,6 +1,6 @@
 package seedu.academydirectory.logic.commands;
 
-import seedu.academydirectory.model.Model;
+import seedu.academydirectory.model.VersionedModel;
 
 /**
  * Terminates the program.
@@ -17,7 +17,7 @@ public class ExitCommand extends Command {
     public static final String MESSAGE_EXIT_ACKNOWLEDGEMENT = "Exiting Academy Directory as requested ...";
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(VersionedModel model) {
         return new CommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT, false, true);
     }
