+++ src/main/java/seedu/academydirectory/logic/commands/GradeCommand.java
@@ -11,6 +11,7 @@ import seedu.academydirectory.commons.core.Messages;
 import seedu.academydirectory.commons.core.index.Index;
 import seedu.academydirectory.logic.commands.exceptions.CommandException;
 import seedu.academydirectory.model.Model;
+import seedu.academydirectory.model.VersionedModel;
 import seedu.academydirectory.model.student.Assessment;
 import seedu.academydirectory.model.student.Student;
 
@@ -62,7 +63,7 @@ public class GradeCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(Model model) throws CommandException {
+    public CommandResult execute(VersionedModel model) throws CommandException {
         requireNonNull(model);
         List<Student> lastShownList = model.getFilteredStudentList();
 
@@ -79,6 +80,7 @@ public class GradeCommand extends Command {
                 assessmentToEdit, studentToEdit.getTags());
         model.setStudent(studentToEdit, editedStudent);
         model.updateFilteredStudentList(Model.PREDICATE_SHOW_ALL_STUDENTS);
+        model.commit(String.format(MESSAGE_SUCCESS, editedStudent.getName(), assessment));
         return new CommandResult(String.format(MESSAGE_SUCCESS, editedStudent.getName(), assessment));
     }
