+++ src/main/java/seedu/academydirectory/logic/commands/HistoryCommand.java
@@ -1,16 +1,14 @@
-package seedu.academydirectory.versioncontrol.logic.commands;
+package seedu.academydirectory.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 
-import seedu.academydirectory.logic.commands.CommandResult;
 import seedu.academydirectory.logic.commands.exceptions.CommandException;
-import seedu.academydirectory.versioncontrol.OptionalVersion;
-import seedu.academydirectory.versioncontrol.Version;
+import seedu.academydirectory.model.VersionedModel;
 
 /**
  * Lists all commands used in the academy directory to the user.
  */
-public class HistoryCommand extends VcCommand {
+public class HistoryCommand extends Command {
     public static final String COMMAND_WORD = "history";
 
     public static final String HELP_MESSAGE = "### Listing all history : `history`\n"
@@ -20,8 +18,8 @@ public class HistoryCommand extends VcCommand {
             + "Format: `history`";
 
     @Override
-    public CommandResult execute(OptionalVersion<Version> version) throws CommandException {
-        requireNonNull(version);
-        return new CommandResult(String.join("\n", version.retrieveHistory()));
+    public CommandResult execute(VersionedModel model) throws CommandException {
+        requireNonNull(model);
+        return new CommandResult(String.join("\n", model.retrieveHistory()));
     }
 }
