+++ src/main/java/seedu/academydirectory/logic/commands/ShowCommand.java
@@ -3,7 +3,7 @@ package seedu.academydirectory.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import javafx.collections.ObservableList;
-import seedu.academydirectory.model.Model;
+import seedu.academydirectory.model.VersionedModel;
 import seedu.academydirectory.model.student.Student;
 
 /**
@@ -65,7 +65,7 @@ public class ShowCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(VersionedModel model) {
         requireNonNull(model);
         ObservableList<Student> students = model.getFilteredStudentList();
         String result = displayResult(students, this.assessment);
