+++ src/main/java/seedu/academydirectory/model/VersionedModel.java
@@ -1,4 +1,4 @@
-package seedu.academydirectory.versioncontrol;
+package seedu.academydirectory.model;
 
 import java.io.IOException;
 import java.text.ParseException;
@@ -6,10 +6,7 @@ import java.util.List;
 
 import seedu.academydirectory.versioncontrol.objects.Commit;
 
-/**
- * API of the Version component
- */
-public interface Version {
+public interface VersionedModel extends Model {
     /**
      * Commits a change and tie the change with a given message
      * @param message Message attached to the Commit for a readable explanation
@@ -19,5 +16,4 @@ public interface Version {
     List<String> retrieveHistory();
 
     Commit revert(String fiveCharHash) throws IOException, ParseException;
-
 }
