--- src/main/java/seedu/academydirectory/versioncontrol/logic/parsers/AcademyDirectoryVcCommandParser.java
@@ -1,56 +1,0 @@
-package seedu.academydirectory.versioncontrol.logic.parsers;
-
-import static seedu.academydirectory.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.academydirectory.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
-
-import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import seedu.academydirectory.commons.core.LogsCenter;
-import seedu.academydirectory.logic.commands.HelpCommand;
-import seedu.academydirectory.logic.parser.exceptions.ParseException;
-import seedu.academydirectory.versioncontrol.logic.commands.HistoryCommand;
-import seedu.academydirectory.versioncontrol.logic.commands.RevertCommand;
-import seedu.academydirectory.versioncontrol.logic.commands.VcCommand;
-
-/**
- * Parses user input.
- */
-public class AcademyDirectoryVcCommandParser {
-    /**
-     * Used for initial separation of command word and args.
-     */
-    private static final Pattern BASIC_COMMAND_FORMAT = Pattern.compile("(?<commandWord>\\S+)(?<arguments>.*)");
-
-    /**
-     * Used for logging input parsed.
-     */
-    private final Logger logger = LogsCenter.getLogger(AcademyDirectoryVcCommandParser.class);
-
-    /**
-     * Parses user input into command for execution.
-     *
-     * @param userInput full user input string
-     * @return the command based on the user input
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public VcCommand parseCommand(String userInput) throws ParseException {
-        final Matcher matcher = BASIC_COMMAND_FORMAT.matcher(userInput.trim());
-        if (!matcher.matches()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, HelpCommand.MESSAGE_USAGE));
-        }
-
-        final String commandWord = matcher.group("commandWord");
-        final String arguments = matcher.group("arguments");
-        switch (commandWord) {
-        case HistoryCommand.COMMAND_WORD:
-            return new HistoryCommand();
-        case RevertCommand.COMMAND_WORD:
-            return new RevertCommandParser().parse(arguments);
-        default:
-            throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
-        }
-    }
-
-}
