+++ src/test/java/seedu/academydirectory/logic/commands/ClearCommandTest.java
@@ -6,24 +6,24 @@ import static seedu.academydirectory.testutil.TypicalStudents.getTypicalAcademyD
 import org.junit.jupiter.api.Test;
 
 import seedu.academydirectory.model.AcademyDirectory;
-import seedu.academydirectory.model.Model;
 import seedu.academydirectory.model.ModelManager;
 import seedu.academydirectory.model.UserPrefs;
+import seedu.academydirectory.model.VersionedModel;
 
 public class ClearCommandTest {
 
     @Test
     public void execute_emptyAcademyDirectory_success() {
-        Model model = new ModelManager();
-        Model expectedModel = new ModelManager();
+        VersionedModel model = new ModelManager();
+        VersionedModel expectedModel = new ModelManager();
 
         assertCommandSuccess(new ClearCommand(), model, ClearCommand.MESSAGE_SUCCESS, expectedModel);
     }
 
     @Test
     public void execute_nonEmptyAcademyDirectory_success() {
-        Model model = new ModelManager(getTypicalAcademyDirectory(), new UserPrefs());
-        Model expectedModel = new ModelManager(getTypicalAcademyDirectory(), new UserPrefs());
+        VersionedModel model = new ModelManager(getTypicalAcademyDirectory(), new UserPrefs());
+        VersionedModel expectedModel = new ModelManager(getTypicalAcademyDirectory(), new UserPrefs());
         expectedModel.setAcademyDirectory(new AcademyDirectory());
 
         assertCommandSuccess(new ClearCommand(), model, ClearCommand.MESSAGE_SUCCESS, expectedModel);
